+++ Bardez.Project.SwordOfTheStars.StructureVisualization/XmlDescription/SaveFileOffsetCollector.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Text;
 using System.Xml;
 
-namespace Bardez.Project.SwordOfTheStars.Editor.UI.WinForms
+namespace Bardez.Project.SwordOfTheStars.StructureVisualization.XmlDescription
 {
     public class OffsetNode
     {
